The tests are debuggable in MSVC and are running as expected
