+++ Libs/CMakeLists.txt
@@ -4,4 +4,5 @@ if (NOT APPLE)
    add_subdirectory(Graphics)
 endif()
 add_subdirectory(Networking)
+add_subdirectory(Utils)
 add_subdirectory(XMPP)
