+++ Libs/Configuration/provider.cpp
@@ -1,11 +1,11 @@
 #include <vector>
 #include <Logging/logging.h>
-#include "loader.h"
-#include <gtest/internal/gtest-port.h>
 
-const std::string Phyre::Configuration::Loader::kWho = "[Configuration::Loader]";
+#include "provider.h"
 
-Phyre::Configuration::Loader::Loader(const std::string& phyre_config_path) :
+const std::string Phyre::Configuration::Provider::kWho = "[Configuration::Provider]";
+
+Phyre::Configuration::Provider::Provider(const std::string& phyre_config_path) :
     target_("phyre.json"),
     phyre_config_path_(phyre_config_path),
     phyre_config_directory_(phyre_config_path)
@@ -17,13 +17,13 @@ Phyre::Configuration::Loader::Loader(const std::string& phyre_config_path) :
     LoadConfigurationFromFile(phyre_config_path_);
 }
 
-bool Phyre::Configuration::Loader::LoadConfigurationFromFile(const std::string& path_to_phyre_config) {
+bool Phyre::Configuration::Provider::LoadConfigurationFromFile(const std::string& path_to_phyre_config) {
     boost::property_tree::ptree root;
     boost::property_tree::read_json(path_to_phyre_config, root);
     return ParseConfiguration(root);
 }
 
-Phyre::Configuration::Application* Phyre::Configuration::Loader::LoadApplication(const std::string& application_name) const {
+Phyre::Configuration::Application* Phyre::Configuration::Provider::LoadApplication(const std::string& application_name) const {
     ApplicationMap::const_iterator cit = application_map_.find(application_name);
     ApplicationMap::const_iterator cend = application_map_.cend();
     if (cit == cend) {
@@ -32,7 +32,7 @@ Phyre::Configuration::Application* Phyre::Configuration::Loader::LoadApplication
     return new Application(cit->second);
 }
 
-std::string Phyre::Configuration::Loader::GetContents(const std::string& target_application, const std::string& resource) const {
+std::string Phyre::Configuration::Provider::GetContents(const std::string& target_application, const std::string& resource) const {
     std::unique_ptr<Application> p_application(LoadApplication(target_application));
     if (p_application) {
         std::ifstream resource_stream = p_application->OpenResource(resource);
@@ -45,7 +45,7 @@ std::string Phyre::Configuration::Loader::GetContents(const std::string& target_
     return std::string();
 }
 
-bool Phyre::Configuration::Loader::ParseConfiguration(boost::property_tree::ptree& pt) {
+bool Phyre::Configuration::Provider::ParseConfiguration(boost::property_tree::ptree& pt) {
     for (const boost::property_tree::ptree::value_type& value : pt.get_child("Phyre.Applications")) {
         for (const boost::property_tree::ptree::value_type& application : value.second) {
             const std::string &name(application.first);
@@ -56,7 +56,7 @@ bool Phyre::Configuration::Loader::ParseConfiguration(boost::property_tree::ptre
     return true;
 }
 
-Phyre::Configuration::Application Phyre::Configuration::Loader::ApplicationFromPtree(const boost::property_tree::ptree::value_type& node) const {
+Phyre::Configuration::Application Phyre::Configuration::Provider::ApplicationFromPtree(const boost::property_tree::ptree::value_type& node) const {
     std::ostringstream full_path;
     full_path << phyre_config_directory_ << node.second.get<std::string>("Path");
     Application::ResourceSet resources;
