+++ Libs/Configuration/provider.h
@@ -15,45 +15,65 @@
 namespace Phyre {
 namespace Configuration {
 
-/*
- * Interface to access the data from the phyre.json configuration structure
- */
-class Loader {
+ /**
+  * \brief Interface to access the data from the phyre.json configuration structure
+  */
+class Provider {
 public:
+    //----------------- Type Definitions --------------
+
+    typedef std::shared_ptr<Provider> Pointer;
+
     //----------------- Construction ------------------
 
-    // @param phyre_config_path The absolute path to phyre.json
-    explicit Loader(const std::string& phyre_config_path);
+    /**
+     * \param phyre_config_path The absolute path to phyre.json 
+     */
+    explicit Provider(const std::string& phyre_config_path);
 
     //----------------- Interface ---------------------
 
-    // The lifetime is expected to be managed by the caller. 
-    // @return Pointer to the loaded application on success. 
-    // @return nullptr on failure.
+    /**
+     * \brief The lifetime is expected to be managed by the caller.
+     * \param application_name 
+     * \return Pointer to the loaded application on success. 
+     * \return nullptr on failure.
+     */
     Application* LoadApplication(const std::string& application_name) const;
 
-    // @param target_application the application from which we are loading the resource data from
-    // @param resource the file we wish to open
-    // @return The contents of the resource. The contents will be empty if there is no resource.
+    /**
+     * \param target_application the application from which we are loading the resource data from
+     * \param resource the file we wish to open
+     * \return The contents of the resource. The contents will be empty if there is no resource.
+     */
     std::string GetContents(const std::string& target_application, const std::string& resource) const;
 
 private:
     //----------------- Type Definitions ---------------------
-    // Map applications by their name
+   
+    /**
+     * \brief Map applications by their name
+     */
     typedef std::unordered_map<std::string, Application> ApplicationMap;
 
     //--------------------- Helpers --------------------------
-    // @param path_to_phyre_config The file we wish to load the configuration from
-    // @return true if the phyre config file was properly loaded into memory
+
+    /**
+     * \param path_to_phyre_config The file we wish to load the configuration from
+     * \return true if the phyre config file was properly loaded into memory
+     */
     bool LoadConfigurationFromFile(const std::string& path_to_phyre_config);
 
-    // @param root The root of the property tree
-    // @return true if the configuration was properly parsed
+    /**
+     * \param root The root of the property tree
+     * \return true if the configuration was properly parsed
+     */
     bool ParseConfiguration(boost::property_tree::ptree& root);
 
     Application ApplicationFromPtree(const boost::property_tree::ptree::value_type& node) const;
 
     //----------------- Data Members ----------------
+    
     // The name of the target phyre.json
     std::string target_;
