+++ Libs/Networking/CMakeLists.txt
@@ -6,5 +6,4 @@ add_library(Networking ${NETWORKING_SRC})
 target_link_libraries(Networking ${Boost_LIBRARIES} Logging)
 set_target_properties(Networking PROPERTIES ARCHIVE_OUTPUT_DIRECTORY ${PHYRE_LIBS}
                                             LIBRARY_OUTPUT_DIRECTORY ${PHYRE_LIBS})
-
 add_subdirectory(Test)
