+++ Libs/Networking/Test/CMakeLists.txt
@@ -3,7 +3,6 @@ cmake_minimum_required(VERSION 3.6)
 include_directories(${PHYRE_INCLUDE_DIRS})
 link_directories(${Boost_LIBRARY_DIRS})
 file(GLOB NETWORKING_TESTS_SRC "*.cpp" "*.h")
-add_library(NetworkingTests ${NETWORKING_TESTS_SRC})
+add_executable(NetworkingTests ${NETWORKING_TESTS_SRC})
 target_link_libraries(NetworkingTests ${Boost_LIBRARIES} gmock gtest Networking)
-set_target_properties(NetworkingTests PROPERTIES ARCHIVE_OUTPUT_DIRECTORY ${PHYRE_LIBS}
-                                                 LIBRARY_OUTPUT_DIRECTORY ${PHYRE_LIBS})
+set_target_properties(NetworkingTests PROPERTIES RUNTIME_OUTPUT_DIRECTORY ${PHYRE_RUNTIME})
