+++ Libs/Networking/Test/networking_tests.h
@@ -1,10 +1,80 @@
 #pragma once
 
+#include <gtest/gtest.h>
+#include <boost/assign.hpp>
+#include <boost/lexical_cast.hpp>
+#include <Logging/logging.h>
+#include <Networking/tcp_client.h>
+#include <Networking/tcp_server.h>
+#include "fake_tcp_clients.h"
+
 namespace Phyre {
 namespace Networking {
-namespace Tests {
-    // Invoke this function to link the configuration tests
-    void Load();
-}
+
+    class TCPClientTest : public ::testing::Test {
+    protected:
+        TCPClientTest():
+            host_("127.0.0.1"),
+            port_or_service_("1234"),
+            tcp_server_(TCPServer(io_service_, boost::lexical_cast<uint16_t>(port_or_service_))) { }
+
+        virtual ~TCPClientTest() { }
+
+        boost::asio::io_service io_service_;
+        std::string host_;
+        std::string port_or_service_;
+        TCPServer tcp_server_;
+    };
+
+    TEST_F(TCPClientTest, CanResolveHosts) {
+        TCPClientHostResolved client(io_service_);
+        tcp_server_.StartAccept();
+        client.Connect(host_, port_or_service_);
+        io_service_.run();
+    }
+
+    TEST_F(TCPClientTest, CanConnectToAService) {
+        TCPClientConnect client(io_service_);
+        tcp_server_.StartAccept();
+        client.Connect(host_, port_or_service_);
+        io_service_.run();
+    }
+
+    TEST_F(TCPClientTest, CanReadAndWriteFromAService) {
+        std::string payload = "Hello!\r\n";
+        TCPClientRead client(io_service_, payload);
+        tcp_server_.StartAccept();
+        client.Connect(host_, port_or_service_);
+        io_service_.run();
+    }
+
+    TEST_F(TCPClientTest, CanDisconnectFromAService) {
+        TCPClientDisconnect client(io_service_);
+        tcp_server_.StartAccept();
+        client.Connect(host_, port_or_service_);
+        io_service_.run();
+    }
+
+    TEST_F(TCPClientTest, DisconnectsAfterError) {
+        TCPClientError client(io_service_);
+
+        // There is no server to connect to, so an error is handled
+        client.Connect(host_, port_or_service_);
+        io_service_.run();
+    }
+
+    TEST_F(TCPClientTest, ServerSendsQueuedMessages) {
+        std::queue<std::string> message_queue;
+        boost::assign::push(message_queue)("first")("second")("third")("fourth")("fith");
+        std::string port_or_service = "1235";
+
+        tcp_server_ = TCPServer(io_service_, boost::lexical_cast<uint16_t>(port_or_service), message_queue);
+        TCPClientReadQueuedMessages client(io_service_, message_queue);
+
+        tcp_server_.StartAccept();
+        client.Connect(host_, port_or_service);
+        io_service_.run();
+    }
+
 }
 }
